Fixed timer

+++ Frequency.pde
@@ -1,37 +1,36 @@
 class Frequency {
-  
-  double period; // indica el intervalo de tiempo entre beat y beat
-  int note; // valor que representa uno de los doce semitonos existentes
-  int octave; // valor que indica la octava de la nota en cuestión
-  
-  Frequency(int n, int o){
+
+  float period; // indica el intervalo de tiempo entre beat y beat
+  float note; // valor que representa uno de los doce semitonos existentes
+  float octave; // valor que indica la octava de la nota en cuestión
+
+  Frequency(float n, float o) {
     setNote(n);
     setOctave(o);
     updatePeriod();
   }
 
-  void setNote(int n){
+  void setNote(float n) {
     note=n;
   }
 
-  void setOctave(int o){
+  void setOctave(float o) {
     octave=o;
   }
 
-  void updatePeriod(){
+  void updatePeriod() {
     period=1/(440*exp(((octave-4)+((note-10)/12))*log(2))); // fórmula que permite sacar el periodo dada cualquier nota (como valor) y octava
   }
 
-  int getNote(){
+  float getNote() {
     return note;
   }
 
-  int getOctave(){
+  float getOctave() {
     return octave;
   }
 
-  double getPeriod(){
+  float getPeriod() {
     return period;
   }
-
 }
